+++ bobgui/cpp/bobgui.hpp
@@ -54,3 +54,4 @@ namespace cpp {
 
 #endif // BOBGUI_CPP_BOBGUI_HPP#include "media.hpp"
 #include "network.hpp"
+#include "system.hpp"